+++ client/src/App.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./App.css";
 import { ApolloProvider } from "@apollo/react-hooks";
 import client from "./ApolloClient";
-import UserContextProvider from "components/UserContextProvider";
+import UserContextProvider from "components-app/UserContextProvider";
 import Routes from "Routes";
 
 const App = () => {
