+++ client/src/components-app/AuthenticatedRoute/index.jsx
